+++ tests/unit/deployments/cloudbuild_tests.py
@@ -14,6 +14,7 @@
 
 # pylint: disable=line-too-long
 # pylint: disable=missing-module-docstring
+# pylint: disable=too-many-positional-arguments
 
 try:
     from importlib.resources import files as import_files